+++ .github/workflows/super-linter.yml
@@ -54,5 +54,5 @@ jobs:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           VALIDATE_HTML: false
 
-          FILTER_REGEX_EXCLUDE: docker/.jupyter/jupyter_notebook_config.py
+          FILTER_REGEX_EXCLUDE: docker/.jupyter/jupyter_lab_config.py
           VALIDATE_SQLFLUFF: false
